added .depend file and added configuration functions in g2b.py

+++ g2b.py
@@ -3,7 +3,6 @@
 import dropbox
 import os.path
 import time
-from git import Repo
 from db_interface import DropboxConnect, DropboxUpload, DropboxDownload
 from tar_interface import make_tarfile, extract
 import tarfile
@@ -11,6 +10,7 @@ import argparse
 import sh
 import shutil
 import hashlib
+import json
 # Get your app key and secret from the Dropbox developer website
 app_key = 'zmuktk74z6ansu7'
 app_secret = '9c7vwi5kgmcbo6i'
@@ -20,12 +20,6 @@ app_secret = '9c7vwi5kgmcbo6i'
 #connect to dropbox -- app validation etc
 client = DropboxConnect(app_key, app_secret)
 
-
-#Out folders:
-cloud_path = "/project/archive.tar.gz"
-localrepo_path = os.path.abspath("mycode")
-tarball = "tarcode.tar.gz"
-tempfolder = os.path.abspath("temp/mycode")
 currentRevision = None
 
 
@@ -54,6 +48,38 @@ def push():
         push();
 #myrepo = Repo(localrepo_path)
 
+def initialize():
+    local_dir=input('Please kind Fucker, give the local repository directory: ')
+    while  not (os.path.isdir(local_dir)):
+        local_dir=input('Dont fuck with me matey, write again')
+
+    cloud_path=input('Please kind Fucker, give the cloud path')
+    cloud_archive_name=input('Please kind Fucker, give the cloud tar.fz name')
+
+
+    configdata={
+            "local":{
+                "repo":local_dir,
+                "tarball":"temp.tar.gz",
+                "temp":"temp",
+            },
+            "cloud":{
+                "path" : cloud_path,
+                "tarball" : cloud_archive_name,
+            }
+        }
+    with open('g2b.config', 'w') as jsonout:
+        json.dump(configdata, jsonout)
+    return configdata
+
+def configure():
+    try:
+        with open('g2b.config') as jsonin:
+            configdata=json.load(jsonin)
+    except Exception:
+        configdata=initialize()
+    finally:
+        return configdata
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--pull", action="store_true")
@@ -63,7 +89,9 @@ if(args.pull):
 	pull()
 if(args.push):
 	push()
-shutil.rmtree('temp')
-os.remove(tarball)
 
-
+config = configure()
+print (config)
+if(os.path.exists("temp")):
+    shutil.rmtree('temp')
+    os.remove(tarball)
